+++ tests/configs/defaultIpWhiteList.js
@@ -1,0 +1,24 @@
+/* Magic Mirror Test config sample ipWhitelist by default
+ *
+ * By Rodrigo Ramírez Norambuena https://rodrigoramirez.com
+ * MIT Licensed.
+ */
+
+var config = {
+	port: 8080,
+
+	language: "en",
+	timeFormat: 24,
+	units: "metric",
+	electronOptions: {
+		webPreferences: {
+			nodeIntegration: true,
+		},
+	},
+
+	modules: [
+	]
+};
+
+/*************** DO NOT EDIT THE LINE BELOW ***************/
+if (typeof module !== "undefined") {module.exports = config;}